+++ src/index.ts
@@ -27,7 +27,7 @@ export {ProgressIndicatorModule} from './modules/progressIndicator.module';
 export {SERVER_BASE_URL, SERVER_AUTH_HEADER, SERVER_COOKIE_HEADER, SCROLL_MAGIC_CONTROLLER} from './types/tokens';
 export {CookieService} from './services/cookies/cookies.service';
 export {StatusCodeService} from './services/statusCode/statusCode.service';
-export {IgnoredInterceptorsService} from './services/ignoredInterceptors/ignoredInterceptors.service';
+export {IgnoredInterceptorsService, IgnoredInterceptorId, HttpRequestIgnoredInterceptorId} from './services/ignoredInterceptors/ignoredInterceptors.service';
 export {NgComponentOutletEx} from "./directives/ngComponentOutletEx/ngComponentOutletEx.directive";
 export {APP_STABLE, APP_STABLE_PROVIDER, extractAppStableResolve} from "./utils/utils.common";
 export * from './misc';